Issue #113: Show a `payment_method` radio-group options inline (by option node id, will be work for all `payment_method` ragio groups).
Explanaiton for the used solution:

It's impossible to stylize it using class attribute in `dds_registration/forms.py`.

This code:

```
RadioSelect(attrs={"class": "XXX"})
```

will produce class attribute for the hidden inner `input`, not to the wrapper, like that:

```
<input class="XXX" ... type="radio" value="INVOICE">
```

TODO: Didn't we want to use icons here? (I'm not shure, but probably it'll be a bit tricky here.)